Develop j

+++ src/main/java/com/whydigit/efit/repo/BasicDetailRepo.java
@@ -10,7 +10,9 @@ import org.springframework.stereotype.Repository;
 import com.whydigit.efit.entity.BasicDetailVO;
 
 @Repository
-public interface BasicDetailRepo extends JpaRepository<BasicDetailVO, Integer> {
+public interface BasicDetailRepo extends JpaRepository<BasicDetailVO, Long> {
 	@Query(value = "select a from BasicDetailVO a Where a.orgId=?1")
 	List<BasicDetailVO> getAllBasicDetailByOrgId(Long orgId);
+
+	
 }
