+++ src/main/java/com/whydigit/efit/repo/LogisticsRepo.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import com.whydigit.efit.entity.LogisticsVO;
 
 @Repository
-public interface LogisticsRepo extends JpaRepository<LogisticsVO, Integer> {
+public interface LogisticsRepo extends JpaRepository<LogisticsVO, Long> {
 	@Query(value = "select a from LogisticsVO a Where a.orgId=?1")
 	List<LogisticsVO> getLogisticsByOrgId(Long orgId);
 }
