+++ src/main/java/com/whydigit/efit/repo/PackingDetailRepo.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import com.whydigit.efit.entity.PackingDetailVO;
 
 @Repository
-public interface PackingDetailRepo extends JpaRepository<PackingDetailVO, Integer> {
+public interface PackingDetailRepo extends JpaRepository<PackingDetailVO, Long> {
 	@Query(value = "select a from PackingDetailVO a Where a.orgId=?1")
 	List<PackingDetailVO> getpackingDetailByOrgId(Long orgId);
 }
