+++ src/main/java/com/whydigit/efit/repo/StockDetailRepo.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import com.whydigit.efit.entity.StockDetailVO;
 
 @Repository
-public interface StockDetailRepo extends JpaRepository<StockDetailVO, Integer> {
+public interface StockDetailRepo extends JpaRepository<StockDetailVO, Long> {
 	@Query(value = "select a from StockDetailVO a Where a.orgId=?1")
 	List<StockDetailVO> getStockDetailByOrgId(Long orgId);
 }
