+++ src/main/java/com/whydigit/efit/service/PartStudyService.java
@@ -12,42 +12,42 @@ public interface PartStudyService {
 
 	List<BasicDetailVO> getAllBasicDetail(Long orgId);
 
-	Optional<BasicDetailVO> getBasicDetailById(int id);
+	Optional<BasicDetailVO> getBasicDetailById(Long id);
 
 	BasicDetailVO createBasicDetail(BasicDetailVO basicDetailVO);
 
 	Optional<BasicDetailVO> updateBasicDetail(BasicDetailVO basicDetailVO);
 
-	void deleteBasicDetail(int id);
+	void deleteBasicDetail(Long id);
 
 	List<PackingDetailVO> getAllpackingDetail(Long orgId);
 
-	Optional<PackingDetailVO> getPackingDetailById(int id);
+	Optional<PackingDetailVO> getPackingDetailById(Long id);
 
 	PackingDetailVO createPackingDetail(PackingDetailVO packingDetailVO);
 
 	Optional<PackingDetailVO> updatePackingDetail(PackingDetailVO packingDetailVO);
 
-	void deletePackingDetail(int id);
+	void deletePackingDetail(Long id);
 
 	List<LogisticsVO> getAllLogistics(Long orgId);
 
-	Optional<LogisticsVO> getLogisticsById(int id);
+	Optional<LogisticsVO> getLogisticsById(Long id);
 
 	LogisticsVO createLogistics(LogisticsVO logisticsVO);
 
 	Optional<LogisticsVO> updateLogistics(LogisticsVO logisticsVO);
 
-	void deleteLogistics(int id);
+	void deleteLogistics(Long id);
 
 	List<StockDetailVO> getAllStockDetail(Long orgId);
 
-	Optional<StockDetailVO> getStockDetailById(int id);
+	Optional<StockDetailVO> getStockDetailById(Long id);
 
 	StockDetailVO createStockDetail(StockDetailVO stockDetailVO);
 
 	Optional<StockDetailVO> updateStockDetail(StockDetailVO stockDetailVO);
 
-	void deleteStockDetail(int id);
+	void deleteStockDetail(Long id);
 
 }
